Decouple Lock from ZMUserSession
