+++ Source/UserSession/ZMUserSession+AppLock.swift
@@ -18,12 +18,59 @@
 
 import Foundation
 
+public protocol UserSessionAppLockInterface {
+
+    var appLockController: AppLockType { get set }
+
+    /// The current session lock, if any.
+
+    var lock: SessionLock? { get }
+
+}
+
+public extension UserSessionAppLockInterface {
+
+    /// Whether the session is currently locked.
+
+    var isLocked: Bool {
+        return lock != .none
+    }
+
+}
+
+extension ZMUserSession: UserSessionAppLockInterface {
+
+    public var lock: SessionLock? {
+        if isDatabaseLocked {
+            return .database
+        } else if appLockController.isLocked {
+            return .screen
+        } else {
+            return nil
+        }
+    }
+
+}
+
+/// The various types of session locks, i.e reasons why the session is inaccessible.
+
+public enum SessionLock {
+
+    /// The session is locked because the has been in the background for too long.
+
+    case screen
+
+    /// The session is locked because the database is locked and not accessible.
+
+    case database
+
+}
+
 protocol UserSessionAppLockDelegate: class {
 
     func userSessionDidUnlock(_ session: ZMUserSession)
 
 }
-
 
 extension ZMUserSession: AppLockDelegate {
