+++ Source/UserSession/ZMUserSession.Configuration.swift
@@ -18,15 +18,6 @@
 
 import Foundation
 
-public protocol UserSessionAppLockInterface {
-
-    var appLockController: AppLockType { get set }
-    var lock: ZMUserSession.Lock? { get }
-
-}
-
-extension ZMUserSession: UserSessionAppLockInterface {}
-
 public extension ZMUserSession {
 
     /// An object used to configure a user session.
@@ -47,4 +38,3 @@ public extension ZMUserSession {
     }
 
 }
-
